Inital sort of working version
Still really just a stub of functionality here.  We have
a basic window scroll sort of effect going, and a crude display.

Still need to flush out some decisions around how to do frames, or how
we might take text and translate it, and how to build a movie by
scrolling or showing frames.